- Deleting DummyCffService and DummyCffServiceImpl - Adding cascade ALL in manyToOne relation of pickupPoint inside AllowedVehicle entity - Changing CffGood id from Long to String - Changing Fleet entity's price data type from float to double, applied to cbmCapacity too - Changing HeaderCff data type from String to Date - Adding cascade ALL inside PickupPoint entity relation with headerCff and merchant - Adding new model : UploadCffAllowedVehicle, UploadCffMerchant, UploadCffPickupPoint, and add them to UploadCffResponse - Add Repository for Category, Fleet, and Warehouse - Add service for Category (get category), Fleet (get fleet), and Warehouse (get warehouse) - Cleaning the JpaCffWriter class. Now writing to merchant table, pickup point table, and allowed_vehicle table. - Creating controller for category, fleet, and warehouse.

+++ indianapolis-model/src/main/java/com/gdn/entity/Fleet.java
@@ -23,10 +23,10 @@ public class Fleet {
     private String name;
 
     @Column(name = "price")
-    private Float price;
+    private Double price;
 
     @Column(name = "cbm_capacity")
-    private Float cbmCapacity;
+    private Double cbmCapacity;
 
     @ManyToOne
     @JoinColumn(name = "logisctic_vendor_id")
